test: Add stencil tests to MPI runners
